+++ core/CMakeLists.txt
@@ -136,7 +136,7 @@ else ()
             SOURCE_FILES ${SOURCE_FILES}
             src/main/c/share/files.c
             src/main/c/share/os.c
-    )
+            src/main/c/share/dedup.cpp)
 endif (WIN32)
 
 set(CMAKE_ARCHIVE_OUTPUT_DIRECTORY ${OUTPUT})
