+++ core/src/main/c/share/util.h
@@ -68,6 +68,12 @@
     #define MM_PREFETCH_NTA(address) _mm_prefetch((address), _MM_HINT_NTA)
 #endif
 
+#ifdef __APPLE__
+#define __JLONG_REINTERPRET_CAST__(type, var)  (type)var
+#else
+#define __JLONG_REINTERPRET_CAST__(type, var)  reinterpret_cast<type>(var)
+#endif
+
 constexpr jdouble D_MAX = std::numeric_limits<jdouble>::infinity();
 constexpr jdouble D_MIN = -std::numeric_limits<jdouble>::infinity();
 constexpr jint I_MAX = std::numeric_limits<jint>::max();
